setup a new reducer component because the reducer can become quite large
