+++ AdvancedReact/react-advanced-v18/src/tutorial/6-useReducer/setup/index.js
@@ -1,37 +1,12 @@
 import React, { useState, useReducer } from 'react';
 import Modal from './Modal';
 import { data } from '../../../data';
-// reducer function
-const reducer = (state, action) => {
-  if (action.type === 'ADD_ITEM') {
-    const newPeople = [...state.people, action.payload];
-    return {
-      ...state,
-      people: newPeople,
-      isModalOpen: true,
-      modalContent: 'item added',
-    };
-  }
-  if (action.type === 'NO_VALUE') {
-    return { ...state, isModalOpen: true, modalContent: 'please enter value' };
-  }
-  if (action.type === 'CLOSE_MODAL') {
-    return { ...state, isModalOpen: false };
-  }
-  if (action.type === 'REMOVE_ITEM') {
-    const newPeople = state.people.filter(
-      (person) => person.id !== action.payload
-    );
-    return { ...state, people: newPeople };
-  }
-  throw new Error('no matching action type');
-};
+import { reducer } from './reducer';
 const defaultState = {
   people: [],
   isModalOpen: false,
   modalContent: 'hello world',
 };
-
 const Index = () => {
   // React Hooks
   const [name, setName] = useState('');
